# swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Skolaroid API
  version: 1.0.0
  description: |
    OpenAPI (Swagger) documentation for Skolaroid API route handlers under `src/app/api`.

    Notes:
    - Many successful responses include additional fields besides `success`, `message`, and `data`.
    - Some endpoints are currently placeholders or transitional implementations.
    - Auth is primarily Supabase session cookie-based unless otherwise noted.
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /api/prisma/group/{groupId}:
    get:
      tags: [Groups]
      summary: Get a private group by id
      operationId: getGroupById
      security:
        - SupabaseSession: []
      parameters:
        - $ref: '#/components/parameters/groupId'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      tags: [Groups]
      summary: Update group settings
      operationId: updateGroup
      security:
        - SupabaseSession: []
      parameters:
        - $ref: '#/components/parameters/groupId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGroupInput'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/prisma/invitation/send:
    post:
      tags: [Invitations]
      summary: Send invitations to a group
      operationId: sendInvitations
      security:
        - SupabaseSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendInvitationsInput'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/prisma/invitation/accept:
    post:
      tags: [Invitations]
      summary: Accept invitation by token
      operationId: acceptInvitation
      security:
        - SupabaseSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvitationActionInput'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '410':
          description: Invitation token expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/prisma/invitation/link:
    post:
      tags: [Invitations]
      summary: Create invitation link token
      operationId: createInvitationLink
      description: Temporary endpoint while invitation system finalization is in progress.
      security:
        - SupabaseSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvitationLinkInput'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    SupabaseSession:
      type: apiKey
      in: cookie
      name: sb-access-token
      description: Supabase access token cookie used for authenticated user session.
    CronBearer:
      type: http
      scheme: bearer
      bearerFormat: CRON_SECRET
      description: Bearer token matching server-side `CRON_SECRET` for cron-only endpoints.
  parameters:
    groupId:
      name: groupId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Private group id
...
  responses:
    Success:
      description: Successful operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SuccessResponse'
    BadRequest:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required or invalid auth
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated but not authorized for this action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Requested resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict with current resource state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          nullable: true
        count:
          type: integer
        deletedCount:
          type: integer
        url:
          type: string
          format: uri
      additionalProperties: true
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          nullable: true
          example: false
        error:
          type: string
        message:
          type: string
        detail:
          type: string
      additionalProperties: true
...
    UpdateGroupInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
        description:
          type: string
          maxLength: 500
        message:
          type: string
          maxLength: 300
...
    SendInvitationsInput:
      type: object
      required:
        - groupId
        - emails
      properties:
        groupId:
          type: string
          format: uuid
        emails:
          type: array
          minItems: 1
          maxItems: 20
          items:
            type: string
            format: email
    InvitationActionInput:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          minLength: 1
    CreateInvitationLinkInput:
      type: object
      required:
        - groupId
      properties:
        groupId:
          type: string
          format: uuid
...